Improvements to the playbook. (#13)
* Skip BPFFS installation when LSB is not available

CentOS does not come with LSB by default. This skips installation of BPFFS when the `ansible_lsb`
fact is not available.

This might not be 100% correct, but it's better than having support only for Ubuntu.

* Install Cilium repository when rendering the template

When attempting to render the helm chart template in a clean helm installation, the playbook fails
because the repository has not yet been installed. This makes sure the repository exists before
attempting to render the template.

Additionally, when there are no repositories installed in helm, the `helm repo list` command exists
with an error. In such cases, the playbook would also fail. Now, errors are ignored when listing the
available repositories and the install command is ran wether the repo command failed or the
repository is not yet available. When the `helm repo list` command throws an error not related to
the absence of repositories, the install command will most likely throw it too, so the play will
fail in such cases.

* Apply suggestions from code review

Co-authored-by: Robert Wimmer <2039811+githubixx@users.noreply.github.com>

Co-authored-by: Robert Wimmer <2039811+githubixx@users.noreply.github.com>